padronizando os dados na API

+++ routes/api.php
@@ -2,45 +2,36 @@
 
 use App\Http\Controllers\AuthController;
 use App\Http\Controllers\BranchController;
-use App\Http\Controllers\DashboardController;
 use App\Http\Controllers\EquipmentController;
 use App\Http\Controllers\UserController;
 use Illuminate\Support\Facades\Route;
 
 // Rotas para autenticação
-Route::prefix('access')->middleware('api')->group(function () {
+Route::prefix('v1')->group(function () {
     Route::post('login', [AuthController::class, 'login'])->name('login');
     Route::post('register', [AuthController::class, 'register']);
-    Route::middleware('auth:api')->get('me', [AuthController::class, 'me']);
-});
 
-Route::prefix('admin')->middleware(['api', 'auth:api', 'role:super_admin|organization_admin'])->group(function () {
-    Route::get('/users', [UserController::class, 'index']);
-    Route::get('/users/{id}', [UserController::class, 'show']);
-    Route::post('/users', [UserController::class, 'store']);
-    Route::put('/users/{id}', [UserController::class, 'update']);
-    Route::delete('/users/{id}', [UserController::class, 'destroy']);
-});
-
-Route::prefix('dashboard')->middleware(['api', 'auth:api'])->group(function () {
-    Route::get('/', [DashboardController::class, 'index']);
-});
-
-Route::prefix('branches')->middleware(['api', 'auth:api'])->group(function () {
-
-    Route::prefix('{branch_id}')->group(function () {
-
-        Route::prefix('equipments')->middleware(['role_or_permission:super_admin|organization_admin|equipment_manager'])->group(function () {
-            Route::get('/', [EquipmentController::class, 'index']);
-            Route::post('/', [EquipmentController::class, 'store']);
-            Route::get('/{equipment}', [EquipmentController::class, 'show']);
-            Route::patch('/{equipment}', [EquipmentController::class, 'update']);
-            Route::delete('/{equipment}', [EquipmentController::class, 'destroy']);
-        });
+    Route::middleware(['auth:api', 'role:super_admin|organization_admin'])->group(function () {
+        Route::get('admin/users', [UserController::class, 'index']);
+        Route::get('admin/users/{id}', [UserController::class, 'show']);
+        Route::post('admin/users', [UserController::class, 'store']);
+        Route::put('admin/users/{id}', [UserController::class, 'update']);
+        Route::delete('admin/users/{id}', [UserController::class, 'destroy']);
     });
 
-    Route::get('/equipments', [BranchController::class, 'getAllEquipments']);
+    Route::middleware(['auth:api'])->group(function () {
+        Route::get('branches/', [BranchController::class, 'index']);
+        Route::prefix('branches/{branch_id}/equipments')->group(function () {
+            Route::get('/', [EquipmentController::class, 'index']);
+            Route::post('/', [EquipmentController::class, 'store']);
+            Route::get('{equipment}', [EquipmentController::class, 'show']);
+            Route::patch('{equipment}', [EquipmentController::class, 'update']);
+            Route::delete('{equipment}', [EquipmentController::class, 'destroy']);
+        });
+    });
 });
+
+
 
 // // Rotas para Histórico de Manutenções
 // Route::prefix('maintenance-history')->middleware(['api', 'auth:api'])->group(function () {
